backend/requirements.txt
# ActionFlow Backend - Dependencies

# ─────────────── Web Framework ───────────────
fastapi>=0.109.0,<0.120.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ─────────────── Database ───────────────
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
alembic>=1.13.0

# ─────────────── LangChain & LangGraph ───────────────
langchain>=0.2.0,<0.4.0
langchain-openai>=0.1.0,<0.3.0
langchain-core>=0.2.0,<0.4.0
langgraph>=0.2.0,<0.3.0

# ─────────────── OpenAI ───────────────
openai>=1.12.0
tiktoken>=0.6.0

# ─────────────── HTTP Client ───────────────
httpx>=0.27.0
aiohttp>=3.9.0
twilio>=8.0.0

# ─────────────── Amadeus API ───────────────
amadeus>=9.0.0

# ─────────────── Data Validation ───────────────
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ─────────────── Environment & Config ───────────────
python-dotenv>=1.0.0

# ─────────────── Utilities ───────────────
typing-extensions>=4.9.0
python-dateutil>=2.8.0

# ─────────────── Redis (Session State) ───────────────
redis>=5.0.0

# ─────────────── Logging & Monitoring ───────────────
structlog>=24.1.0
prometheus-fastapi-instrumentator>=7.0.0

# ─────────────── Testing (optional) ───────────────
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.23.0,<0.24.0
pytest-cov>=4.1.0

# ─────────────── Voice Support ───────────────
assemblyai>=0.23.0
elevenlabs>=1.0.0

pinecone-client==3.0.0
langchain-pinecone==0.1.0

prometheus-client>=0.19.0
email-validator>=2.0.0
